+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Next
 
 - [#9](https://github.com/AshFurrow/ARCollectionViewMasonryLayout/pull/9): Moved ARCollectionViewMasonryLayout.podspec and screenshots into the project - [@dblock](https://github.com/dblock).
 - [#11](https://github.com/AshFurrow/ARCollectionViewMasonryLayout/issues/11): Fixed crash in layout without footer on iOS8 - [@dblock](https://github.com/dblock).
+- [#13](https://github.com/AshFurrow/ARCollectionViewMasonryLayout/pull/13): Added `headerHeight` and `headerViewClass` - [@dblock](https://github.com/dblock).
+- [#13](https://github.com/AshFurrow/ARCollectionViewMasonryLayout/pull/13): Removed `leadingView` and `trailingView` until properly implemented - [@dblock](https://github.com/dblock).
 
 0.0.2
 ----------------